extract components from Textarea
